requirements.txt
gitpython~=3.1.37
rich~=13.3.5